Organize lambdas in sub folders in src
