+++ src/api/get_anime/__init__.py
